add ecloud build fix for mac agnet.yml

+++ build-ecloud.sh
@@ -27,20 +27,14 @@ if [ "$(git symbolic-ref --short HEAD)"=="master" ]; then
 fi
 
 #build
-#make clean config build-linux
-#make config build-darwin
-
-sed -i "s/build-linux-amd64: config/build-linux-amd64:/" $WORKBASE/framework/Makefile 
-sed -i "s/build-darwin: config/build-darwin:/" $WORKBASE/framework/Makefile 
-
-make clean update-vfs update-generated-file update-plugins build-linux-amd64 build-darwin
-
-sed -i "s/build-linux-amd64:/build-linux-amd64: config/" $WORKBASE/framework/Makefile
-sed -i "s/build-darwin:/build-darwin: config/" $WORKBASE/framework/Makefile
+make clean config build-linux
+make config build-darwin
 
 #copy-configs
 cp -rf $WORKBASE/framework/LICENSE $WORKDIR/bin && cat $WORKBASE/framework/NOTICE $WORKDIR/NOTICE > $WORKDIR/bin/NOTICE
-cp -rf $PNAME.yml $WORKDIR/bin
+if [ ! -f $WORKDIR/bin/$PNAME.yml ]; then
+  cp -rf $PNAME.yml $WORKDIR/bin
+fi
 
 cd $WORKDIR/bin
 for t in amd64 ; do
